requirements.txt
PyYAML==3.12
requests==2.18.4
  certifi==2018.1.18
  chardet==3.0.4
  idna==2.6
  urllib3==1.22
six==1.11.0
flake8==3.5.0
   mccabe==0.6.1
   pycodestyle==2.3.1
   pyflakes==1.6.0
tox==3.0.0
  pluggy==0.6.0
  py==1.5.3
  virtualenv==15.2.0
requests-mock==1.4.0
coverage==4.5.1
pip==10.0.1
pylint==1.8.4
  astroid==1.6.3
  isort==4.3.4
  lazy-object-proxy==1.3.1
  wrapt==1.10.11
